% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.dlm.R
\name{build.dlm}
\alias{build.dlm}
\title{Construir objeto DLM}
\usage{
build.dlm(y = rep(0, 5), f = list(X = NULL, irregular = "S", nivel = "S",
  inclinacao = "S", sazon = "S", regres = "S", freq = 4), pre.interv = F)
}
\arguments{
\item{y}{Série de tempo a ser decomposta}

\item{f}{Lista de fundamentos para a construção do objeto dlm:
\itemize{
\item \code{X}: Variáveis independentes
\item \code{irregular}: Tipo de variância do resíduo principal 
\item \code{nivel}: Tipo de variância do resíduo do nível
\item \code{inclinacao}: Tipo de variância do resíduo do tendência
\item \code{sazon}: Tipo de variância do resíduo da sazonalidade
\item \code{regres}: Tipo de variância do resíduo dos coeficientes 
\item \code{freq}: Frequência da Sazonalidade, em geral a mesma que a da série }}

\item{pre.interv}{Parametro indica se o modelo será posteriormente utilizado para
detecção de intervenções. Padrão \code{FALSE}.}
}
\value{
Objeto dlm
}
\description{
Função que cria a partir dos parâmetros desejados as matrizes no formato dlm.
Para mais informações ver pacote dlm: \url{https://cran.r-project.org/package=dlm}.
}
\seealso{
\code{\link{criar.dlm}}
\code{\link{dlmMLE2}}
}
\keyword{internal}

