% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.ssm.R
\name{build.ssm}
\alias{build.ssm}
\title{Construir objeto DLM}
\usage{
build.ssm(y, ipar, f = list(X = NULL, irregular = NA, nivel = NA, inclinacao =
  NA, sazon = NA, regres = NA, freq = 4), pre.interv = F, ...)
}
\arguments{
\item{y}{Série de tempo a ser decomposta}

\item{f}{Lista de fundamentos para a construção do objeto dlm:
\itemize{
\item \code{X}: Variáveis independentes
\item \code{irregular}: Tipo de variância do resíduo principal 
\item \code{nivel}: Tipo de variância do resíduo do nível
\item \code{inclinacao}: Tipo de variância do resíduo do tendência
\item \code{sazon}: Tipo de variância do resíduo da sazonalidade
\item \code{regres}: Tipo de variância do resíduo dos coeficientes 
\item \code{freq}: Frequência da Sazonalidade, em geral a mesma que a da série }}
}
\value{
Objeto dlm
}
\description{
Função que cria a partir dos parâmetros desejados as matrizes no formato dlm.
Para mais informações ver pacote dlm: \url{https://cran.r-project.org/web/packages/dlm/dlm.pdf}.
}
\seealso{
\code{\link{decompor}}
\code{\link{dlmMLE2}}
}
\keyword{internal}
