% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choques.R
\name{choques}
\alias{choques}
\title{Calculo de Choques}
\usage{
choques(kf)
}
\arguments{
\item{kf}{Objeto do tipo dlmFiltered. Ver \code{dlmFilter}}
}
\value{
lista com os choques e seus desvios padrão
  \itemize{
    \item \code{ch} : Matrix com choques com a seguinte ordem: Choque Principal,
      Choque no Nível, Choque na Inclinação
    \item \code{sd} : Matrix com desvios padrão, segue mesma ordem que \code{ch}
  }
}
\description{
Função retorna os choques, ou disturbios, nas equações principais e nas 
equações de transição
}
\keyword{internal}

