% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompor.todos.R
\name{decompor.todos}
\alias{decompor.todos}
\title{Estimação de alternativas hipoteses}
\usage{
decompor.todos(y, X = NULL, comeco = NULL, fim = NULL, sazon.b = TRUE,
  regres = "S")
}
\arguments{
\item{y}{Série de tempo a ser decomposta}

\item{X}{Série de tempo das variáveis independentes}

\item{comeco}{Período inicial dos cálculos}

\item{fim}{Período final dos cálculos}

\item{sazon.b}{Boolean indicativa se o efeito sazonal será incorporado na forma funcional, Padrão = TRUE}

\item{regres}{Define que os coeficientes devem ser fixos no tempo ("F") ou estocásticos ("S")}
}
\value{
lista com os 8 modelos estimados
}
\description{
Função que calcula 8 formas funcionais para elasticidades para fins de comparação
}
\seealso{
\code{\link{decompor}}
\code{\link{exportar}}
}
