% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmMLE2.R
\name{dlmMLE2}
\alias{dlmMLE2}
\title{Otimizacao especializada}
\usage{
dlmMLE2(y, ipar, fun, ...)
}
\arguments{
\item{y}{Série temporal a ser decomposta}

\item{ipar}{Parâmetros iniciais da função de otimização}

\item{fun}{Função de geração do objeto dlm, geralmente build.dlm}

\item{...}{Parâmetros da função de geração do objeto dlm}
}
\value{
Parâmetros ótimos para a função de geração do objeto dlm
}
\description{
Esta função prepara os dados para chamar a função optim
}
\keyword{internal}

