% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar.R
\name{exportar}
\alias{exportar}
\title{Exportacao de resultado}
\usage{
exportar(resultado)
}
\arguments{
\item{resultado}{Lista resultante da função \code{\link{calcularElasticidades}} com o conjunto completo de elasticidades calculadas}
}
\value{
Data frame no formato correto para exportação para o Excel. Em sistemas Windows o data.frame é copiado também
para o clipboard, possibilitando a colagem das informações no Excel com um simples Ctrl-V.
}
\description{
Função que formato ao dados para um padrão de exportação.#'
}
\examples{
\dontrun{data(Exemplo)}
\donttest{resultado <- calcularElasticidades(Exemplo$receitas, Exemplo$Hpib, Exemplo$Hpet, fim=c(2015,4))}

\donttest{resultado.exportacao <- exportar(resultado)}
\donttest{write.csv2(resultado.exportacao, "/tmp/dados.csv")} #Escreve resultados em arquivo CSV.
}
\seealso{
\code{\link{todas.dlms}}
\code{\link{calcularElasticidades}}
\code{\link{Exemplo}}
}
