% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervencoes.R
\name{intervencoes}
\alias{intervencoes}
\title{Intervencoes}
\usage{
intervencoes(kfs)
}
\arguments{
\item{u}{Data frame de choques}

\item{dlm}{Objeto dlm}
}
\value{
Lista com posição da intervenção
}
\description{
Função para detectar automaticamente as intervenções
}
\keyword{internal}
