% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdlm.R
\name{jdlm}
\alias{jdlm}
\title{Modelo em função do tempo}
\usage{
jdlm(dlm, n = NULL)
}
\arguments{
\item{dlm}{Objeto 'dlm'. Ver função dlm}

\item{n}{opcional se dlm$X não for NULL. Número de períodos de série de
tempo principal.}
}
\value{
Lista com componentes do modelo dlm:
  \itemize{
    \item \code{FF} : Lista com n matrizes de transformação. Se dlm$JFF for NULL
      então todas as matrizes da lista são identicas. Se não são incorparados
      os valores corretos de dlm$X.
    \item \code{GG} : Lista com n matrizes de transição. Se dlm$JGG for NULL
      então todas as matrizes da lista são identicas. Se não são incorparados
      os valores corretos de dlm$X.
    \item \code{V} : Lista com n matrizes de variâncias dos choques na equação
      principal. Se dlm$JV for NULL então todas as matrizes da lista são
      identicas. Se não são incorparados os valores corretos de dlm$X.
    \item \code{W} : Lista com n matrizes de variâncias dos choques na equação
      de transição. Se dlm$JW for NULL então todas as matrizes da lista são
      identicas. Se não são incorparados os valores corretos de dlm$X.
 }
}
\description{
Função transforma objetos dlm tradicionais em listas com modelos para 
cada momento no tempo.
}
\keyword{internal}
