% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelar.R
\name{modelar}
\alias{modelar}
\title{Componentes e Testes}
\usage{
modelar(fit)
}
\arguments{
\item{y}{Série temporal a ser decomposta}

\item{dlm}{Objeto 'dlm'}

\item{estrutura}{Lista com variáveis importantes para realização de testes:
\itemize{
  \item \code{p}: Número de variâncias estocásticas
  \item \code{logLike}: Valor da função de log-verossimilhança }}
}
\value{
Lista com 11 variáveis, entre componentes e testes:
  \itemize{
    \item \code{y}: Série que foi decomposta
    \item \code{dlm}: Estrutura do objeto dlm usado na decomposição
    \item \code{f}: Componentes resultantes do filtro
    \item \code{theta}: Componentes suavizados
    \item \code{c}: Subconjunto de theta, com sazonalidade agregada e nomes nas colunas (Resultado Principal)
    \item \code{interv}: tabela listando as intervenções, seu componente, período, valor e teste t.
    \item \code{u}: Matriz de resíduos
    \item \code{q}: Teste de independência dos resíduos
    \item \code{h}: Teste de homocedasticidade dos resíduos
    \item \code{nt}: Teste de normalidade dos resíduos
    \item \code{aic}: Valor AIC da função de verossimilhança usada para estimar modelo.}
}
\description{
Função que calcula a partir do objeto dlm e de algumas variáveis estruturais,
o valor dos componentes, reestrutura o resultados da função do pacote dlm para fins
de legibilidade, e, por fim, realiza testes diversos nos resíduos e valores obtidos.
}
\keyword{internal}
