% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posicoes.R
\name{posicoes}
\alias{posicoes}
\title{Posicoes dos componentes}
\usage{
posicoes(dlm)
}
\arguments{
\item{dlm}{Objeto dlm}
}
\value{
Vetor com as posições dos componentes no objeto dlm. names(ret): nome dos componentes existentes
}
\description{
Função que detecta a partir de um objeto dlm as posições dos componentes
}
\keyword{internal}

