% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.h.R
\name{teste.h}
\alias{teste.h}
\title{Teste H}
\usage{
teste.h(u, d)
}
\arguments{
\item{u}{Matriz de resíduos}

\item{d}{Número de Componentes}
}
\value{
Valor do Test H
}
\description{
Função para testar a homocedasticidade dos resíduos, teste H
}
\keyword{internal}
