% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.normalidade.R
\name{teste.normalidade}
\alias{teste.normalidade}
\title{Teste de normalidade}
\usage{
teste.normalidade(u)
}
\arguments{
\item{u}{Matriz de resíduos}
}
\value{
Valor do teste de normalidade
}
\description{
Função para testar a normalidade dos resíduos
}
\keyword{internal}

