% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Exemplo}
\alias{Exemplo}
\title{Exemplos de Inputs e Outputs}
\format{An object of class \code{list} of length 29.}
\usage{
data(Exemplo)
}
\description{
Lista extensiva de dados utilizados somente para exemplos e testes.
}
\keyword{datasets}

